Start refactoring the threaded implementation
